invoiceproject/requirements.txt
asgiref==3.7.2
Django==5.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.19
phonenumberslite==8.13.21
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.4
